Fix some kinks in the database schema
The database schema's relationship between pipelines and
remotes was a bit awkward.

This fixes that by introducing a new type for a project, and then
letting git remotes just be attached to projects so that a project
can have multiple remotes (same url different branch or even
different urls altogether). Pipelines can then just be attached to
the remotes so that different branches can run the same pipeline.